@@ -16,65 +16,51 @@
 
 #include <benchmarks/common/benchmark_utilities.hpp>
 #include <benchmarks/common/generate_input.hpp>
-#include <benchmarks/common/table_utilities.hpp>
-#include <benchmarks/fixture/benchmark_fixture.hpp>
-#include <benchmarks/synchronization/synchronization.hpp>
+#include <benchmarks/common/nvbench_utilities.hpp>
 
 #include <cudf/column/column_view.hpp>
 #include <cudf/reduction.hpp>
 #include <cudf/types.hpp>
 
+#include <nvbench/nvbench.cuh>
+
 #include <memory>
 
-class Reduction : public cudf::benchmark {};
-
-template <typename type>
-void BM_reduction_anyall(benchmark::State& state,
-                         std::unique_ptr<cudf::reduce_aggregation> const& agg)
+template <typename DataType>
+static void reduction_anyall(nvbench::state& state, nvbench::type_list<DataType>)
 {
-  cudf::size_type const column_size{static_cast<cudf::size_type>(state.range(0))};
-  auto const dtype           = cudf::type_to_id<type>();
-  data_profile const profile = data_profile_builder().no_validity().distribution(
-    dtype, distribution_id::UNIFORM, 0, agg->kind == cudf::aggregation::ANY ? 0 : 100);
-  auto const values = create_random_column(dtype, row_count{column_size}, profile);
+  auto const size     = static_cast<cudf::size_type>(state.get_int64("size"));
+  auto const kind_str = state.get_string("kind");
 
-  cudf::data_type output_dtype{cudf::type_id::BOOL8};
+  auto const input_type = cudf::type_to_id<DataType>();
+  auto const agg        = kind_str == "any" ? cudf::make_any_aggregation<cudf::reduce_aggregation>()
+                                            : cudf::make_all_aggregation<cudf::reduce_aggregation>();
 
-  for (auto _ : state) {
-    cuda_event_timer timer(state, true);
-    auto result = cudf::reduce(*values, *agg, output_dtype);
-  }
+  data_profile const profile =
+    data_profile_builder().no_validity().distribution(input_type,
+                                                      distribution_id::UNIFORM,
+                                                      (kind_str == "all" ? 1 : 0),
+                                                      (kind_str == "any" ? 0 : 100));
+  auto const values = create_random_column(input_type, row_count{size}, profile);
 
-  // The benchmark takes a column and produces one scalar.
-  set_items_processed(state, column_size + 1);
-  set_bytes_processed(state, estimate_size(values->view()) + cudf::size_of(output_dtype));
+  auto const output_type = cudf::data_type{cudf::type_id::BOOL8};
+  auto stream            = cudf::get_default_stream();
+  state.set_cuda_stream(nvbench::make_cuda_stream_view(stream.value()));
+  state.add_element_count(size);
+  state.add_global_memory_reads<DataType>(size);
+  state.add_global_memory_writes<nvbench::int8_t>(1);
+
+  state.exec(nvbench::exec_tag::sync, [&values, output_type, &agg](nvbench::launch& launch) {
+    cudf::reduce(*values, *agg, output_type);
+  });
+
+  set_throughputs(state);
 }
 
-#define concat(a, b, c) a##b##c
-#define get_agg(op)     concat(cudf::make_, op, _aggregation<cudf::reduce_aggregation>())
+using Types = nvbench::type_list<bool, int8_t, int32_t, float>;
 
-// TYPE, OP
-#define RBM_BENCHMARK_DEFINE(name, type, aggregation)             \
-  BENCHMARK_DEFINE_F(Reduction, name)(::benchmark::State & state) \
-  {                                                               \
-    BM_reduction_anyall<type>(state, get_agg(aggregation));       \
-  }                                                               \
-  BENCHMARK_REGISTER_F(Reduction, name)                           \
-    ->UseManualTime()                                             \
-    ->Arg(10000)      /* 10k */                                   \
-    ->Arg(100000)     /* 100k */                                  \
-    ->Arg(1000000)    /* 1M */                                    \
-    ->Arg(10000000)   /* 10M */                                   \
-    ->Arg(100000000); /* 100M */
-
-#define REDUCE_BENCHMARK_DEFINE(type, aggregation) \
-  RBM_BENCHMARK_DEFINE(concat(type, _, aggregation), type, aggregation)
-
-REDUCE_BENCHMARK_DEFINE(bool, all);
-REDUCE_BENCHMARK_DEFINE(int8_t, all);
-REDUCE_BENCHMARK_DEFINE(int32_t, all);
-REDUCE_BENCHMARK_DEFINE(float, all);
-REDUCE_BENCHMARK_DEFINE(bool, any);
-REDUCE_BENCHMARK_DEFINE(int8_t, any);
-REDUCE_BENCHMARK_DEFINE(int32_t, any);
-REDUCE_BENCHMARK_DEFINE(float, any);
+NVBENCH_BENCH_TYPES(reduction_anyall, NVBENCH_TYPE_AXES(Types))
+  .set_name("anyall")
+  .set_type_axes_names({"DataType"})
+  .add_string_axis("kind", {"any", "all"})
+  .add_int64_axis("size", {100'000, 1'000'000, 10'000'000, 100'000'000});
