@@ -16,79 +16,84 @@
 
 #include <benchmarks/common/benchmark_utilities.hpp>
 #include <benchmarks/common/generate_input.hpp>
-#include <benchmarks/fixture/benchmark_fixture.hpp>
-#include <benchmarks/synchronization/synchronization.hpp>
+#include <benchmarks/common/nvbench_utilities.hpp>
 
+#include <cudf/aggregation.hpp>
 #include <cudf/dictionary/encode.hpp>
 #include <cudf/reduction.hpp>
 #include <cudf/types.hpp>
 #include <cudf/unary.hpp>
 
-class ReductionDictionary : public cudf::benchmark {};
+#include <nvbench/nvbench.cuh>
 
-template <typename T>
-void BM_reduction_dictionary(benchmark::State& state,
-                             std::unique_ptr<cudf::reduce_aggregation> const& agg)
+template <cudf::reduce_aggregation::Kind kind>
+static std::unique_ptr<cudf::reduce_aggregation> make_reduce_aggregation()
 {
-  cudf::size_type const column_size{static_cast<cudf::size_type>(state.range(0))};
+  switch (kind) {
+    case cudf::reduce_aggregation::ANY:
+      return cudf::make_any_aggregation<cudf::reduce_aggregation>();
+    case cudf::reduce_aggregation::ALL:
+      return cudf::make_all_aggregation<cudf::reduce_aggregation>();
+    case cudf::reduce_aggregation::MIN:
+      return cudf::make_min_aggregation<cudf::reduce_aggregation>();
+    case cudf::reduce_aggregation::MAX:
+      return cudf::make_max_aggregation<cudf::reduce_aggregation>();
+    case cudf::reduce_aggregation::MEAN:
+      return cudf::make_mean_aggregation<cudf::reduce_aggregation>();
+    default: CUDF_FAIL("Unsupported reduce aggregation in this benchmark");
+  }
+}
 
-  // int column and encoded dictionary column
+template <typename DataType, cudf::reduce_aggregation::Kind kind>
+static void reduction_dictionary(nvbench::state& state,
+                                 nvbench::type_list<DataType, nvbench::enum_type<kind>>)
+{
+  cudf::size_type const size{static_cast<cudf::size_type>(state.get_int64("size"))};
+
   data_profile const profile = data_profile_builder().cardinality(0).no_validity().distribution(
     cudf::type_to_id<long>(),
     distribution_id::UNIFORM,
-    (agg->kind == cudf::aggregation::ALL ? 1 : 0),
-    (agg->kind == cudf::aggregation::ANY ? 0 : 100));
-  auto int_column = create_random_column(cudf::type_to_id<long>(), row_count{column_size}, profile);
-  auto number_col = cudf::cast(*int_column, cudf::data_type{cudf::type_to_id<T>()});
+    (kind == cudf::aggregation::ALL ? 1 : 0),
+    (kind == cudf::aggregation::ANY ? 0 : 100));
+  auto int_column = create_random_column(cudf::type_to_id<long>(), row_count{size}, profile);
+  auto number_col = cudf::cast(*int_column, cudf::data_type{cudf::type_to_id<DataType>()});
   auto values     = cudf::dictionary::encode(*number_col);
 
-  cudf::data_type output_dtype = [&] {
-    if (agg->kind == cudf::aggregation::ANY || agg->kind == cudf::aggregation::ALL)
+  cudf::data_type output_type = [&] {
+    if (kind == cudf::aggregation::ANY || kind == cudf::aggregation::ALL) {
       return cudf::data_type{cudf::type_id::BOOL8};
-    if (agg->kind == cudf::aggregation::MEAN) return cudf::data_type{cudf::type_id::FLOAT64};
-    return cudf::data_type{cudf::type_to_id<T>()};
+    }
+    if (kind == cudf::aggregation::MEAN) { return cudf::data_type{cudf::type_id::FLOAT64}; }
+    return cudf::data_type{cudf::type_to_id<DataType>()};
   }();
 
-  for (auto _ : state) {
-    cuda_event_timer timer(state, true);
-    auto result = cudf::reduce(*values, *agg, output_dtype);
+  auto agg = make_reduce_aggregation<kind>();
+
+  auto stream = cudf::get_default_stream();
+  state.set_cuda_stream(nvbench::make_cuda_stream_view(stream.value()));
+  state.add_element_count(size);
+  state.add_global_memory_reads<DataType>(size);
+  if (kind == cudf::aggregation::ANY || kind == cudf::aggregation::ALL) {
+    state.add_global_memory_writes<nvbench::int8_t>(1);  // BOOL8s
+  } else {
+    state.add_global_memory_writes<DataType>(1);
   }
 
-  // The benchmark takes a column and produces two scalars.
-  set_items_processed(state, column_size + 1);
+  state.exec(nvbench::exec_tag::sync, [&values, output_type, &agg](nvbench::launch& launch) {
+    cudf::reduce(*values, *agg, output_type);
+  });
 
-  // We don't set the metrics for the size read/written as row_bit_count() doesn't
-  // support the dictionary type yet (and so is estimate_size()).
-  // See https://github.com/rapidsai/cudf/issues/16121 for details.
+  set_throughputs(state);
 }
 
-#define concat(a, b, c) a##b##c
-#define get_agg(op)     concat(cudf::make_, op, _aggregation<cudf::reduce_aggregation>())
+using Types    = nvbench::type_list<int32_t, float>;
+using AggKinds = nvbench::enum_type_list<cudf::reduce_aggregation::ALL,
+                                         cudf::reduce_aggregation::ANY,
+                                         cudf::reduce_aggregation::MIN,
+                                         cudf::reduce_aggregation::MAX,
+                                         cudf::reduce_aggregation::MEAN>;
 
-// TYPE, OP
-#define RBM_BENCHMARK_DEFINE(name, type, aggregation)                       \
-  BENCHMARK_DEFINE_F(ReductionDictionary, name)(::benchmark::State & state) \
-  {                                                                         \
-    BM_reduction_dictionary<type>(state, get_agg(aggregation));             \
-  }                                                                         \
-  BENCHMARK_REGISTER_F(ReductionDictionary, name)                           \
-    ->UseManualTime()                                                       \
-    ->Arg(10000)      /* 10k */                                             \
-    ->Arg(100000)     /* 100k */                                            \
-    ->Arg(1000000)    /* 1M */                                              \
-    ->Arg(10000000)   /* 10M */                                             \
-    ->Arg(100000000); /* 100M */
-
-#define REDUCE_BENCHMARK_DEFINE(type, aggregation) \
-  RBM_BENCHMARK_DEFINE(concat(type, _, aggregation), type, aggregation)
-
-REDUCE_BENCHMARK_DEFINE(int32_t, all);
-REDUCE_BENCHMARK_DEFINE(float, all);
-REDUCE_BENCHMARK_DEFINE(int32_t, any);
-REDUCE_BENCHMARK_DEFINE(float, any);
-REDUCE_BENCHMARK_DEFINE(int32_t, min);
-REDUCE_BENCHMARK_DEFINE(float, min);
-REDUCE_BENCHMARK_DEFINE(int32_t, max);
-REDUCE_BENCHMARK_DEFINE(float, max);
-REDUCE_BENCHMARK_DEFINE(int32_t, mean);
-REDUCE_BENCHMARK_DEFINE(float, mean);
+NVBENCH_BENCH_TYPES(reduction_dictionary, NVBENCH_TYPE_AXES(Types, AggKinds))
+  .set_name("reduction_dictionary")
+  .set_type_axes_names({"DataType", "AggKinds"})
+  .add_int64_axis("size", {100'000, 1'000'000, 10'000'000, 100'000'000});
