@@ -16,9 +16,7 @@
 
 #include <benchmarks/common/benchmark_utilities.hpp>
 #include <benchmarks/common/generate_input.hpp>
-#include <benchmarks/common/table_utilities.hpp>
-#include <benchmarks/fixture/benchmark_fixture.hpp>
-#include <benchmarks/synchronization/synchronization.hpp>
+#include <benchmarks/common/nvbench_utilities.hpp>
 
 #include <cudf/column/column.hpp>
 #include <cudf/column/column_view.hpp>
@@ -26,43 +24,38 @@
 #include <cudf/table/table.hpp>
 #include <cudf/types.hpp>
 
-class ReductionScan : public cudf::benchmark {};
+#include <nvbench/nvbench.cuh>
 
-template <typename type>
-static void BM_reduction_scan(benchmark::State& state, bool include_nulls)
+template <typename DataType>
+static void reduction_scan(nvbench::state& state, nvbench::type_list<DataType>)
 {
-  cudf::size_type const n_rows{(cudf::size_type)state.range(0)};
-  auto const dtype  = cudf::type_to_id<type>();
-  auto const column = create_random_column(dtype, row_count{n_rows});
-  if (!include_nulls) column->set_null_mask(rmm::device_buffer{}, 0);
+  auto const size       = static_cast<cudf::size_type>(state.get_int64("size"));
+  auto const nulls      = state.get_float64("nulls");
+  auto const input_type = cudf::type_to_id<DataType>();
 
-  std::unique_ptr<cudf::column> result = nullptr;
-  for (auto _ : state) {
-    cuda_event_timer timer(state, true);
-    result = cudf::scan(
-      *column, *cudf::make_min_aggregation<cudf::scan_aggregation>(), cudf::scan_type::INCLUSIVE);
-  }
+  data_profile const profile = data_profile_builder().null_probability(nulls).distribution(
+    input_type, distribution_id::UNIFORM, 0, 100);
+  auto const input_column = create_random_column(input_type, row_count{size}, profile);
 
-  // The benchmark takes a column and produces a new column of the same size as input.
-  set_items_processed(state, n_rows * 2);
-  set_bytes_processed(state, estimate_size(column->view()) + estimate_size(result->view()));
+  auto agg = cudf::make_min_aggregation<cudf::scan_aggregation>();
+
+  auto stream = cudf::get_default_stream();
+  state.set_cuda_stream(nvbench::make_cuda_stream_view(stream.value()));
+  state.add_element_count(size);
+  state.add_global_memory_reads<DataType>(size);
+  state.add_global_memory_writes<DataType>(1);
+
+  state.exec(nvbench::exec_tag::sync, [&input_column, &agg](nvbench::launch& launch) {
+    cudf::scan(*input_column, *agg, cudf::scan_type::INCLUSIVE);
+  });
+
+  set_throughputs(state);
 }
 
-#define SCAN_BENCHMARK_DEFINE(name, type, nulls)                          \
-  BENCHMARK_DEFINE_F(ReductionScan, name)                                 \
-  (::benchmark::State & state) { BM_reduction_scan<type>(state, nulls); } \
-  BENCHMARK_REGISTER_F(ReductionScan, name)                               \
-    ->UseManualTime()                                                     \
-    ->Arg(10000)      /* 10k */                                           \
-    ->Arg(100000)     /* 100k */                                          \
-    ->Arg(1000000)    /* 1M */                                            \
-    ->Arg(10000000)   /* 10M */                                           \
-    ->Arg(100000000); /* 100M */
+using Types = nvbench::type_list<int8_t, int32_t, uint64_t, float, int16_t, uint32_t, double>;
 
-SCAN_BENCHMARK_DEFINE(int8_no_nulls, int8_t, false);
-SCAN_BENCHMARK_DEFINE(int32_no_nulls, int32_t, false);
-SCAN_BENCHMARK_DEFINE(uint64_no_nulls, uint64_t, false);
-SCAN_BENCHMARK_DEFINE(float_no_nulls, float, false);
-SCAN_BENCHMARK_DEFINE(int16_nulls, int16_t, true);
-SCAN_BENCHMARK_DEFINE(uint32_nulls, uint32_t, true);
-SCAN_BENCHMARK_DEFINE(double_nulls, double, true);
+NVBENCH_BENCH_TYPES(reduction_scan, NVBENCH_TYPE_AXES(Types))
+  .set_name("scan")
+  .set_type_axes_names({"DataType"})
+  .add_float64_axis("nulls", {0.0, 0.1})
+  .add_int64_axis("size", {100'000, 1'000'000, 10'000'000, 100'000'000});
